refactor: InputHandler 추상화
