+++ src/main/java/cleancode/studycafe/jiihyun/io/ConsoleInputHandler.java
@@ -1,0 +1,44 @@
+package cleancode.studycafe.jiihyun.io;
+
+import cleancode.studycafe.jiihyun.exception.AppException;
+import cleancode.studycafe.jiihyun.model.StudyCafePass;
+import cleancode.studycafe.jiihyun.model.StudyCafePassType;
+
+import java.util.List;
+import java.util.Scanner;
+
+public class ConsoleInputHandler implements InputHandler {
+
+    private static final Scanner SCANNER = new Scanner(System.in);
+
+    @Override
+    public StudyCafePassType getPassTypeSelectingUserAction() {
+        String userInput = SCANNER.nextLine();
+
+        if ("1".equals(userInput)) {
+            return StudyCafePassType.HOURLY;
+        }
+        if ("2".equals(userInput)) {
+            return StudyCafePassType.WEEKLY;
+        }
+        if ("3".equals(userInput)) {
+            return StudyCafePassType.FIXED;
+        }
+        throw new AppException("잘못된 입력입니다.");
+    }
+
+    @Override
+    public StudyCafePass getSelectPass(List<StudyCafePass> passes) {
+        String userInput = SCANNER.nextLine();
+        int selectedIndex = Integer.parseInt(userInput) - 1;
+        return passes.get(selectedIndex);
+    }
+
+    @Override
+    public boolean getLockerSelection() {
+        String userInput = SCANNER.nextLine();
+        return "1".equals(userInput);
+    }
+
+}
+
