+++ src/main/java/cleancode/studycafe/jiihyun/io/InputHandler.java
@@ -1,40 +1,16 @@
 package cleancode.studycafe.jiihyun.io;
 
-import cleancode.studycafe.jiihyun.exception.AppException;
 import cleancode.studycafe.jiihyun.model.StudyCafePass;
 import cleancode.studycafe.jiihyun.model.StudyCafePassType;
 
 import java.util.List;
-import java.util.Scanner;
 
-public class InputHandler {
+public interface InputHandler {
 
-    private static final Scanner SCANNER = new Scanner(System.in);
+    StudyCafePassType getPassTypeSelectingUserAction();
 
-    public StudyCafePassType getPassTypeSelectingUserAction() {
-        String userInput = SCANNER.nextLine();
+    StudyCafePass getSelectPass(List<StudyCafePass> passes);
 
-        if ("1".equals(userInput)) {
-            return StudyCafePassType.HOURLY;
-        }
-        if ("2".equals(userInput)) {
-            return StudyCafePassType.WEEKLY;
-        }
-        if ("3".equals(userInput)) {
-            return StudyCafePassType.FIXED;
-        }
-        throw new AppException("잘못된 입력입니다.");
-    }
-
-    public StudyCafePass getSelectPass(List<StudyCafePass> passes) {
-        String userInput = SCANNER.nextLine();
-        int selectedIndex = Integer.parseInt(userInput) - 1;
-        return passes.get(selectedIndex);
-    }
-
-    public boolean getLockerSelection() {
-        String userInput = SCANNER.nextLine();
-        return "1".equals(userInput);
-    }
+    boolean getLockerSelection();
 
 }
